requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
Flask>=3.0.0
duckdb>=1.0.0
PyYAML>=6.0.0
pymongo>=4.6.0